ci(clippy): extend the gate to test/bench code (--all-targets)
Follows up the lib+bins Clippy gate (#18) by linting test, bench and
example code too. Cleared the ~20 findings --all-targets surfaced, all in
inline #[cfg(test)] modules and benches:

- Mechanical: manual_range_contains, byte_char_slices, manual_abs_diff,
  same_item_push, useless_vec, for_kv_map, int_plus_one, unnecessary_cast
  (mix of clippy --fix and hand-fixes).
- eq_op (deny-by-default) in the BAM reg2bin test helper: clippy const-
  evaluates ((1 << 3) - 1) / 7 as 7/7. That is the correct SAM-spec
  bin-offset formula ((1 << 3k) - 1) / 7, kept for consistency with the
  other levels; targeted #[allow(clippy::eq_op)] with a note.
- One unnecessary_cast fix (bayesian test) removed a cast that was pinning
  a float type; re-pinned via "let x = 2.0_f64" instead.

The earlier-noted erasing_op in cyanea-ml/hmm.rs tests was investigated
and is a non-issue: emission[0 * 6 + k] is intentional row-major index
clarity, and current stable clippy does not flag it.

CI: the Clippy job now runs "cargo clippy --workspace --all-targets --
-D warnings". Verified: clippy clean (stable, --all-targets), fmt clean,
"cargo test --workspace --exclude cyanea-py" = 3653 passed, 0 failed.

+++ cyanea-io/src/abi.rs
@@ -412,7 +412,7 @@ mod tests {
         for ch in 0..4u16 {
             let ch_offset = data9_offset + (ch as usize) * trace_bytes;
             for i in 0..trace_len {
-                let value = ((ch as i16 + 1) * 100 + i as i16) as i16;
+                let value = (ch as i16 + 1) * 100 + i as i16;
                 let byte_offset = ch_offset + i * 2;
                 buf[byte_offset..byte_offset + 2].copy_from_slice(&value.to_be_bytes());
             }
